+++ packages/components/src/Button/components/GenericButton.tsx
@@ -1,16 +1,15 @@
 import React, {
+  ComponentType,
   forwardRef,
   Ref,
   useImperativeHandle,
   useRef,
-  ComponentType,
-  FocusEvent,
   MouseEvent,
+  FocusEvent,
 } from "react"
 import classnames from "classnames"
 import { Badge, BadgeAnimated } from "@kaizen/draft-badge"
 import { LoadingSpinner } from "@kaizen/loading-spinner"
-import { ButtonProps } from "../Button"
 import styles from "./GenericButton.module.scss"
 
 export type CustomButtonProps = {
@@ -60,14 +59,6 @@ export type WorkingUndefinedProps = {
   working?: false
 }
 
-type Props = ButtonProps & {
-  additionalContent?: React.ReactNode
-  iconButton?: boolean
-  directionalLink?: boolean
-  paginationLink?: boolean
-  isActive?: boolean
-}
-
 export type BadgeProps = {
   text: string
   animateChange?: boolean
@@ -75,7 +66,32 @@ export type BadgeProps = {
   reversed?: boolean
 }
 
+export type RenderProps = ButtonProps & {
+  additionalContent?: React.ReactNode
+  iconButton?: boolean
+  directionalLink?: boolean
+  paginationLink?: boolean
+  isActive?: boolean
+}
+
 export type ButtonRef = { focus: () => void }
+
+export type ButtonProps = GenericProps &
+  ButtonFormAttributes &
+  (WorkingProps | WorkingUndefinedProps) & {
+    label: string
+    primary?: boolean
+    destructive?: boolean
+    secondary?: boolean
+    /** @default "regular" */
+    size?: "small" | "regular"
+    badge?: BadgeProps
+    type?: "submit" | "reset" | "button"
+    fullWidth?: boolean
+    iconPosition?: "start" | "end"
+    icon?: JSX.Element
+    disabled?: boolean
+  }
 
 // We're treating custom props as anything that is kebab cased.
 // This is so we can support properties like aria-* or data-*
@@ -88,7 +104,7 @@ const getCustomProps = (props: Record<string, any>): Record<string, string> => {
 }
 
 const GenericButton = forwardRef(
-  (props: Props, ref: Ref<ButtonRef | undefined>) => {
+  (props: RenderProps, ref: Ref<ButtonRef | undefined>) => {
     const buttonRef = useRef<HTMLButtonElement | HTMLAnchorElement>()
     useImperativeHandle(ref, () => ({
       focus: (): void => {
@@ -135,7 +151,7 @@ GenericButton.displayName = "GenericButton"
 
 const renderCustomComponent = (
   CustomComponent: ComponentType<CustomButtonProps>,
-  props: Props
+  props: RenderProps
 ): JSX.Element => {
   const { id, disabled, href, onClick, onFocus, onBlur, ...rest } = props
   const customProps = getCustomProps(rest)
@@ -157,7 +173,7 @@ const renderCustomComponent = (
 }
 
 const renderButton = (
-  props: Props,
+  props: RenderProps,
   ref: Ref<HTMLButtonElement>
 ): JSX.Element => {
   const {
@@ -211,7 +227,10 @@ const renderButton = (
   )
 }
 
-const renderLink = (props: Props, ref: Ref<HTMLAnchorElement>): JSX.Element => {
+const renderLink = (
+  props: RenderProps,
+  ref: Ref<HTMLAnchorElement>
+): JSX.Element => {
   const {
     id,
     href,
@@ -246,7 +265,7 @@ const renderLink = (props: Props, ref: Ref<HTMLAnchorElement>): JSX.Element => {
   )
 }
 
-const buttonClass = (props: Props): string => {
+const buttonClass = (props: RenderProps): string => {
   const isDefault = !props.primary && !props.destructive && !props.secondary
   return classnames(
     styles.button,
@@ -273,7 +292,7 @@ const renderLoadingSpinner = (): JSX.Element => (
 )
 
 const renderWorkingContent = (
-  props: Extract<Props, { working: true }>
+  props: Extract<RenderProps, { working: true }>
 ): JSX.Element => {
   if (props.workingLabelHidden) {
     return (
@@ -303,7 +322,7 @@ const renderWorkingContent = (
   )
 }
 
-const renderDefaultContent = (props: Props): JSX.Element => (
+const renderDefaultContent = (props: RenderProps): JSX.Element => (
   <>
     {props.icon && props.iconPosition !== "end" && renderIcon(props.icon)}
     {(!props.icon || !props.iconButton) && (
@@ -319,7 +338,7 @@ const renderDefaultContent = (props: Props): JSX.Element => (
   </>
 )
 
-const renderBadge = (props: Props): JSX.Element | null => {
+const renderBadge = (props: RenderProps): JSX.Element | null => {
   if (!props.badge) return null
 
   const { text, animateChange, reversed, variant } = props.badge
@@ -338,7 +357,7 @@ const renderBadge = (props: Props): JSX.Element | null => {
   )
 }
 
-const renderContent = (props: Props): JSX.Element => (
+const renderContent = (props: RenderProps): JSX.Element => (
   <span className={styles.content}>
     {props.working ? renderWorkingContent(props) : renderDefaultContent(props)}
   </span>
@@ -351,7 +370,7 @@ const renderIcon = (icon: JSX.Element): JSX.Element => (
 // We only want an aria-label in the case that the button has just an icon and no text
 // This can happen when the button is working and workingLabelHidden is true,
 // or when this is an IconButton
-const generateAriaLabel = (props: Props): string | undefined => {
+const generateAriaLabel = (props: RenderProps): string | undefined => {
   if (props.working && props.workingLabelHidden) {
     return props.workingLabel
   }