+++ packages/components/src/FilterBar/context/reducer/setupFilterBarState.ts
@@ -1,8 +1,10 @@
 import { Filters, FiltersValues } from "../../types"
 import { FilterBarState } from "../types"
+import { updateDependentFilters } from "../utils/updateDependentFilters"
 
 export const setupFilterBarState = <ValuesMap extends FiltersValues>(
-  filters: Filters<ValuesMap>
+  filters: Filters<ValuesMap>,
+  values: Partial<ValuesMap>
 ): FilterBarState<ValuesMap> => {
   const state = filters.reduce<FilterBarState<ValuesMap>>(
     (baseState, { id, name, isRemovable, isUsableWhen }) => {
@@ -14,28 +16,27 @@ export const setupFilterBarState = <ValuesMap extends FiltersValues>(
         isRemovable: isRemovable ?? false,
         isUsableWhen,
         isOpen: false,
-        // A dependent filter is set to `null` here as it
-        // will be re-evaluated in the `update_values` dispatch action.
-        isUsable: hasDependency ? null : true,
+        isUsable: true,
       }
 
       if (hasDependency) {
         baseState.dependentFilterIds.add(id)
       }
 
+      if (!isRemovable || values[id] !== undefined) {
+        baseState.activeFilterIds.add(id)
+      }
+
       return baseState
     },
     {
       filters: {},
-      // These will be set by the `update_values` dispatch action.
       activeFilterIds: new Set(),
-      // To prevent an infinite loop calculating dependent filters,
-      // `values` is set to `null` and default values will be
-      // set by the `update_values` dispatch action.
-      values: null,
+      values,
       dependentFilterIds: new Set(),
+      hasUpdatedValues: false,
     } as FilterBarState<ValuesMap>
   )
 
-  return state
+  return updateDependentFilters(state)
 }