+++ packages/components/src/FilterSelect/_docs/FilterSelect.stories.tsx
@@ -1,6 +1,7 @@
 import React, { useState } from "react"
 import { Meta, StoryFn } from "@storybook/react"
 import { renderTriggerControls } from "~components/Filter/_docs/controls/renderTriggerControls"
+import { classNameOverrideArgType } from "../../../../../storybook/argTypes"
 import { FilterButton } from "../../FilterButton"
 import { FilterSelect } from "../FilterSelect"
 import { SelectOption } from "../types"
@@ -10,6 +11,7 @@ export default {
   title: "Components/Filter Select",
   component: FilterSelect,
   argTypes: {
+    ...classNameOverrideArgType,
     isOpen: { control: "disabled" },
     setIsOpen: { control: "disabled" },
     renderTrigger: renderTriggerControls,