+++ packages/date-picker/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Change Log
 
+## 6.2.5
+
+### Patch Changes
+
+- [#3823](https://github.com/cultureamp/kaizen-design-system/pull/3823) [`f924fdeed`](https://github.com/cultureamp/kaizen-design-system/commit/f924fdeed3a83c46cb64ac880cbcc72b5bc6b5e8) - Update TimeField react-aria deps
+
+- [#3831](https://github.com/cultureamp/kaizen-design-system/pull/3831) [`5ad6c577c`](https://github.com/cultureamp/kaizen-design-system/commit/5ad6c577c7f2826ea7e285b0b8d1629643ce7b16) - Update `react-aria` dependencies.
+
 ## 6.2.4
 
 ### Patch Changes